feat: Setup redux store and reducer - Create a new function root reducer function to combine all reducers - Move redux store to the root directory of the project - Modify filepath of redux store in files using it
