+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/serviceWorker.js":"1","/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/features/counter/Counter.js":"2"},{"size":5141,"mtime":1607811333660,"results":"3","hashOfConfig":"4"},{"size":1626,"mtime":1607811333657,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1k7gi9k",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/serviceWorker.js",[],"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/features/counter/Counter.js",[]]
+[{"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/serviceWorker.js":"1","/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/features/counter/Counter.js":"2","/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/store.js":"3","/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/features/counter/counterSlice.js":"4"},{"size":5141,"mtime":1607811333660,"results":"5","hashOfConfig":"6"},{"size":1626,"mtime":1607811333657,"results":"7","hashOfConfig":"6"},{"size":227,"mtime":1607816869307,"results":"8","hashOfConfig":"6"},{"size":1551,"mtime":1607811333658,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1k7gi9k",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/serviceWorker.js",[],"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/features/counter/Counter.js",[],"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/store.js",[],"/Users/doyinolarewaju/Projects/Personal/construyo/frontend-task/src/features/counter/counterSlice.js",[]]
