+++ src/root-reducer.js
@@ -1,0 +1,6 @@
+import { combineReducers } from '@reduxjs/toolkit'
+import counterReducer from "./features/counter/counterSlice";
+
+export default combineReducers({
+    counter: counterReducer
+})
