+++ src/store.js
@@ -1,0 +1,6 @@
+import { configureStore } from '@reduxjs/toolkit';
+import rootReducer from './root-reducer';
+
+export default configureStore({
+  reducer: rootReducer
+});
